Сделал декомпозицию элементов, добавил пропсы
